Crate pathfinder_geometry

Basic geometry and linear algebra primitives, optimized with SIMD.

Modules

angle
Angle utilities.

line_segment
Line segment types, optimized with SIMD.

rect
2D axis-aligned rectangles, optimized with SIMD.

transform2d
2D affine transforms.

transform3d
3D transforms that can be applied to paths.

unit_vector
A utility module that allows unit vectors to be treated like angles.

util
Various utilities.

vector
A SIMD-optimized point type.