Struct pathfinder_geometry::transform2d::Transform2F

pub struct Transform2F {
    pub matrix: Matrix2x2F,
    pub vector: Vector2F,
}

An affine transform, optimized with SIMD.

Fields
matrix: Matrix2x2Fvector: Vector2F
Implementations

impl Transform2F

pub fn from_scale<S>(scale: S) -> Transform2F where
S: IntoVector2F,

pub fn from_rotation(theta: f32) -> Transform2F

pub fn from_rotation_vector(vector: UnitVector) -> Transform2F

pub fn from_translation(vector: Vector2F) -> Transform2F

pub fn from_scale_rotation_translation<S>(
scale: S,
theta: f32,
translation: Vector2F
) -> Transform2F where
S: IntoVector2F,

pub fn row_major(
m11: f32,
m12: f32,
m21: f32,
m22: f32,
m31: f32,
m32: f32
) -> Transform2F

pub fn to_3d(&self) -> Transform4F

pub fn is_identity(&self) -> bool

pub fn extract_scale(&self) -> Vector2F

Extracts the scale from this matrix.

pub fn m11(&self) -> f32

pub fn m21(&self) -> f32

pub fn m12(&self) -> f32

pub fn m22(&self) -> f32

pub fn m31(&self) -> f32

pub fn m32(&self) -> f32

pub fn translate(&self, vector: Vector2F) -> Transform2F

pub fn rotate(&self, theta: f32) -> Transform2F

pub fn scale<S>(&self, scale: S) -> Transform2F where
S: IntoVector2F,

pub fn translation(&self) -> Vector2F

Returns the translation part of this matrix.

This decomposition assumes that scale, rotation, and translation are applied in that order.

pub fn rotation(&self) -> f32

Returns the rotation angle of this matrix.

This decomposition assumes that scale, rotation, and translation are applied in that order.

pub fn scale_factor(&self) -> f32

Returns the scale factor of this matrix.

This decomposition assumes that scale, rotation, and translation are applied in that order.

pub fn inverse(&self) -> Transform2F