Struct pathfinder_geometry::vector::Vector2F

pub struct Vector2F(pub F32x2);

2D points with 32-bit floating point coordinates.

Tuple Fields
0: F32x2
Implementations

impl Vector2F

pub fn new(x: f32, y: f32) -> Vector2F

pub fn splat(value: f32) -> Vector2F

pub fn zero() -> Vector2F

pub fn to_3d(self) -> Vector3F

pub fn to_4d(self) -> Vector4F

pub fn x(self) -> f32

pub fn y(self) -> f32

pub fn set_x(&mut self, x: f32)

pub fn set_y(&mut self, y: f32)

pub fn min(self, other: Vector2F) -> Vector2F

pub fn max(self, other: Vector2F) -> Vector2F

pub fn clamp(self, min_val: Vector2F, max_val: Vector2F) -> Vector2F

pub fn det(self, other: Vector2F) -> f32

pub fn dot(self, other: Vector2F) -> f32

pub fn floor(self) -> Vector2F

pub fn ceil(self) -> Vector2F

pub fn round(self) -> Vector2F

Rounds both coordinates to the nearest integer.

pub fn square_length(self) -> f32

Treats this point as a vector and calculates its squared length.

pub fn length(self) -> f32

Treats this point as a vector and calculates its length.

pub fn normalize(self) -> Vector2F

Treats this point as a vector and normalizes it.

pub fn yx(self) -> Vector2F

Swaps y and x.

pub fn projection_coefficient(self, a: Vector2F) -> f32

Returns the coefficient when the given vector a is projected onto this one.

That is, if this vector is v and this function returns c, then proj_v a = cv. In other words, this function computes (a⋅v) / (v⋅v).

pub fn is_zero(self) -> bool

pub fn lerp(self, other: Vector2F, t: f32) -> Vector2F

pub fn to_i32(self) -> Vector2I